+++ src/components/LogoLockup/index.jsx
@@ -21,9 +21,7 @@ const LogoLockup = ({ className }) => (
   />
 )
 
-LogoLockup.propTypes = {
-  className: string
-}
+LogoLockup.propTypes = { className: string }
 
 export default LogoLockup
 export { LogoLockup }